pyproject.toml
[tool.poetry]
name = "eda-ru-parser"
version = "0.1.0"
description = "A utility to get recipies from eda.ru"
authors = ["Konstantin Meshcheryakov <kmeshcheryakov@klika-tech.com>"]
readme = "README.md"
packages = [{include = "eda_ru_parser"}]

[tool.poetry.dependencies]
python = ">=3.8, <4"
requests = "^2.28.2"
beautifulsoup4 = "^4.12.0"
click = "^8.1.3"
pillow = "^9.4.0"
flask = "^2.3.2"
flask-api = "^3.0.post1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.1.1"

[tool.poetry.scripts]
recipe_downloader = "eda_ru_parser.main:parse_recipe"
recipe_server = "eda_ru_parser.server:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
